requirements.txt
Cython==3.0.1
numpy==1.26.4
absl-py==1.4.0
accelerate==1.0.1
aiobotocore==2.7.0
aiohttp==3.8.5
aioitertools==0.11.0
aiosignal==1.3.1
albumentations==1.3.1
antialiased-cnns==0.3
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
artifactory==0.1.17
astroid==2.15.6
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
boto3==1.28.35
botocore==1.31.35
braceexpand==0.1.7
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
clean-fid==0.1.6
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cmake==3.27.2
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.1.0
coverage==7.3.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==4.4.2
deepspeed==0.10.1
defusedxml==0.7.1
diffusers @ git+https://github.com/huggingface/diffusers.git@2764db3194fc1b5069df7292fd938657d8568995
dill==0.3.7
dm-tree==0.1.8
docker-pycreds==0.4.0
dohq-artifactory==0.9.0
dominate==2.8.0
duckdb==0.10.2
easydict==1.13
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.1.3
executing==2.0.1
ExifRead==3.0.0
faiss-cpu==1.9.0
fastapi==0.103.0
fastjsonschema==2.19.1
fastrlock==0.8.1
filelock==3.12.2
flake8==6.1.0
flash-attn==2.6.3
flatbuffers==24.3.25
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2024.5.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
g5nie==0.1.1
gast==0.5.4
gdown==4.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h11==0.14.0
h5py==3.12.1
hjson==3.1.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.26.1
humanfriendly==10.0
hypothesis==6.102.4
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
iniconfig==2.0.0
insightface==0.7.3
iopath==0.1.10
ipykernel==6.29.3
ipython==8.22.2
isoduration==20.11.0
isort==5.12.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.22
jsonnet==0.20.0
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.1
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
keras==3.6.0
kiwisolver==1.4.5
kornia==0.7.0
lazy-object-proxy==1.9.0
lazy_loader==0.3
libclang==16.0.6
lightning-utilities==0.9.0
lit==16.0.6
lmdb==1.4.1
lpips==0.1.4
lz4==4.3.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
moviepy==1.0.3
mpld3==0.5.9
mpmath==1.3.0
multidict==6.0.4
namex==0.0.7
natsort==8.2.0
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.2
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
notebook_shim==0.2.4
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-cupti-cu12==12.2.142
nvidia-cuda-nvrtc-cu12==12.2.140
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu12==11.0.8.103
nvidia-curand-cu12==10.3.3.141
nvidia-cusolver-cu12==11.5.2.141
nvidia-cusparse-cu12==12.1.2.141
nvidia-nccl-cu12==2.18.3
nvidia-nvtx-cu12==12.2.140
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
open-clip-torch==2.22.0
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
opt-einsum==3.3.0
optree==0.13.0
orjson==3.10.10
overrides==7.7.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pathlib==1.0.1
pathtools==0.1.2
pexpect==4.9.0
Pillow==10.0.0
Pillow-SIMD==9.0.0.post1
platformdirs==3.10.0
pluggy==1.2.0
portalocker==2.8.2
prettytable==3.11.0
proglog==0.1.10
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-spy==0.3.14
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pydantic==1.10.13
pyDeprecate==0.3.2
pyflakes==3.1.0
Pygments==2.17.2
PyJWT==2.8.0
pylint==2.17.5
pynvml==11.5.0
pypandoc==1.11
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.0.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qoi==0.3.1
qudida==0.0.4
referencing==0.33.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rotary-embedding-torch==0.2.7
rpds-py==0.18.0
rsa==4.9
s3fs==2024.5.0
s3transfer==0.6.2
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.13.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
smmap==5.0.0
snappy==3.1.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlglot==23.13.1
sqlglotrs==0.2.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.18.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.3.0
terminado==0.18.1
tf_keras==2.18.0
threadpoolctl==3.2.0
tifffile==2023.8.12
timm==0.9.7
tinycss2==1.2.1
tokenizers==0.20.1
tomli==2.0.1
tomlkit==0.12.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
transformers==4.45.0
triton @ git+https://github.com/openai/triton@e6216047b8b0aef1fe8da6ca8667a3ad0a016411#subdirectory=python
typer==0.9.4
types-python-dateutil==2.8.19.20240311
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
uvicorn==0.23.2
vision-aided-loss==0.1.0
wandb==0.15.8
wcwidth==0.2.6
webcolors==1.13
webdataset==0.2.31
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.3.7
wrapt==1.15.0
xformers==0.0.22+a24a6e3.d20231010
xxhash==3.3.0
yacs==0.1.8
yarl==1.9.2
zipp==3.16.2
zstandard==0.21.0
zstd==1.5.5.1
